+++ app/src/main/java/com/example/yp_playlist_maker/TrackResponse.kt
@@ -1,6 +1,6 @@
 package com.example.yp_playlist_maker
 
-import com.example.yp_playlist_maker.domain.models.track.Track
+import com.example.yp_playlist_maker.domain.models.Track
 
 class TrackResponse(
     val results: ArrayList<Track>
