+++ app/src/main/java/com/example/yp_playlist_maker/data/dto/TrackDto.kt
@@ -1,0 +1,13 @@
+package com.example.yp_playlist_maker.data.dto
+
+data class TrackDto(
+    var trackName: String,
+    var artistName: String,
+    val trackTimeMillis: String,
+    val artworkUrl100: String,
+    val collectionName: String,
+    val releaseDate: String,
+    val primaryGenreName: String,
+    val country: String,
+    val previewUrl: String
+)
