--- app/src/main/java/com/example/yp_playlist_maker/data/dto/app_theme/AppThemeParamsDto.kt
@@ -1,5 +1,0 @@
-package com.example.yp_playlist_maker.data.dto.app_theme
-
-data class AppThemeParamsDto(
-    val switcherStatus: Boolean
-)
