+++ app/src/main/java/com/example/yp_playlist_maker/domain/api/app_theme/AppThemeRepository.kt
@@ -1,6 +1,6 @@
 package com.example.yp_playlist_maker.domain.api.app_theme
 
-import com.example.yp_playlist_maker.domain.models.app_theme.AppThemeParams
+import com.example.yp_playlist_maker.domain.models.AppThemeParams
 
 interface AppThemeRepository {
