+++ app/src/main/java/com/example/yp_playlist_maker/domain/impl/AppThemeInteractorImpl.kt
@@ -1,8 +1,8 @@
-package com.example.yp_playlist_maker.domain.impl.app_theme
+package com.example.yp_playlist_maker.domain.impl
 
 import com.example.yp_playlist_maker.domain.api.app_theme.AppThemeInteractor
 import com.example.yp_playlist_maker.domain.api.app_theme.AppThemeRepository
-import com.example.yp_playlist_maker.domain.models.app_theme.AppThemeParams
+import com.example.yp_playlist_maker.domain.models.AppThemeParams
 
 class AppThemeInteractorImpl(private val repository: AppThemeRepository): AppThemeInteractor {
