+++ app/src/main/java/com/example/yp_playlist_maker/domain/models/AppThemeParams.kt
@@ -1,4 +1,4 @@
-package com.example.yp_playlist_maker.domain.models.app_theme
+package com.example.yp_playlist_maker.domain.models
 
 data class AppThemeParams (
     val switcherStatus: Boolean
