+++ app/src/main/java/com/example/yp_playlist_maker/domain/models/Track.kt
@@ -1,4 +1,4 @@
-package com.example.yp_playlist_maker.domain.models.track
+package com.example.yp_playlist_maker.domain.models
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
