+++ app/src/main/java/com/example/yp_playlist_maker/presentation/settings_activity/SettingsActivity.kt
@@ -10,7 +10,7 @@ import android.widget.Switch
 import androidx.appcompat.app.AppCompatActivity
 import com.example.yp_playlist_maker.Creator
 import com.example.yp_playlist_maker.R
-import com.example.yp_playlist_maker.domain.models.app_theme.AppThemeParams
+import com.example.yp_playlist_maker.domain.models.AppThemeParams
 import com.example.yp_playlist_maker.presentation.application.App
 
 class SettingsActivity : AppCompatActivity() {
